Thread pool affinity support on Solaris/Linux.
This is a PR#9 from yvoinov on Mar 4, 2018 which still was
not merged.
---
This is affinity support on Solaris/Linux. It is well-tested and
increasing performance/decreasing latency of pool for
similar tasks execution. To enable it just define AFFINITY
(-DAFFINITY). Threads binds across online CPU-cores using
Round-Robin. In my case, this reduces pool latency ~ three
times.

Also all int replaced to std::size_t due to portability reasons.